refactor: remove _getGossipPeers from pubsub
